+++ internal/generator/exhaustiveswitch/doc.go
@@ -1,3 +1,3 @@
-// Package exhaustiveswitch generates methods that perform an exhaustive switch
-// over finite set types such as "one-of groups" and enumerations.
+// Package exhaustiveswitch generates functions that perform an exhaustive
+// switch over finite set types such as "one-of groups" and enumerations.
 package exhaustiveswitch