+++ internal/generator/exhaustiveswitch/oneof.go
@@ -7,34 +7,125 @@ import (
 	"github.com/dogmatiq/primo/internal/generator/internal/scope"
 )
 
-// generateForOneOf generates a switch function for the options within a one-of
-// group.
 func generateForOneOf(code *jen.File, g *scope.OneOfGroup) {
-	name := "Switch_" + g.Message.GoTypeName + "_" + g.GoFieldName
-	const paramPrefix = "case"
+	generateOneOfSwitch(code, g)
+	generateOneOfMap(code, g)
+}
+
+func oneOfCaseFuncName(o *scope.OneOfOption) string {
+	return "case" + o.DiscriminatorFieldName
+}
+
+func generateOneOfSwitch(code *jen.File, g *scope.OneOfGroup) {
+	funcName := "Switch_" + g.Message.GoTypeName + "_" + g.GoFieldName
 
 	code.
 		Commentf(
-			"%s dispatches to one of the given functions based on",
-			name,
+			"%s invokes one of the given functions based on",
+			funcName,
 		)
 	code.
 		Commentf(
-			"which value of the [%s] message's %q one-of group is populated.",
-			g.Message.GoTypeName,
+			"the value of x.%s.",
 			g.GoFieldName,
 		)
 	code.Comment("")
 	code.Commentf(
-		"It panics if x.%s field is nil; otherwise, it returns the value",
+		"It panics if x.%s is nil.",
 		g.GoFieldName,
 	)
-	code.Comment("returned by the called function. If no return value is required, use a return")
-	code.Comment("type of [error] and always return nil.")
 
 	code.
 		Func().
-		Id(name).
+		Id(funcName).
+		ParamsFunc(
+			func(code *jen.Group) {
+				code.
+					Line().
+					Id("x").
+					Op("*").
+					Id(g.Message.GoTypeName)
+
+				for _, o := range g.Options {
+					code.
+						Line().
+						Id(oneOfCaseFuncName(o)).
+						Func().
+						Params(
+							o.Field.GoType(),
+						)
+				}
+
+				code.Line()
+			},
+		).
+		Block(
+			jen.
+				Switch(
+					jen.
+						Id("v").
+						Op(":=").
+						Id("x").
+						Dot(g.GoFieldName).
+						Op(".").
+						Call(jen.Type()),
+				).
+				BlockFunc(
+					func(code *jen.Group) {
+						for _, o := range g.Options {
+							code.
+								Case(
+									jen.
+										Op("*").
+										Id(o.DiscriminatorTypeName),
+								).
+								Id(oneOfCaseFuncName(o)).
+								Call(
+									jen.
+										Id("v").
+										Dot(o.DiscriminatorFieldName),
+								)
+						}
+
+						code.
+							Default().
+							Panic(
+								jen.Lit(
+									fmt.Sprintf(
+										"%s: x.%s is nil",
+										funcName,
+										g.GoFieldName,
+									),
+								),
+							)
+					},
+				),
+		)
+}
+
+func generateOneOfMap(code *jen.File, g *scope.OneOfGroup) {
+	funcName := "Map_" + g.Message.GoTypeName + "_" + g.GoFieldName
+
+	code.
+		Commentf(
+			"%s maps x.%s to a value of type T by invoking",
+			funcName,
+			g.GoFieldName,
+		)
+	code.Comment("one of the given functions.")
+	code.Comment("")
+	code.Commentf(
+		"It invokes the function that corresponds to the value of x.%s,",
+		g.GoFieldName,
+	)
+	code.Commentf(
+		"and returns that function's result. It panics if x.%s is nil.",
+		g.GoFieldName,
+	)
+
+	code.
+		Func().
+		Id(funcName).
 		Types(
 			jen.Id("T").Any(),
 		).
@@ -49,7 +140,7 @@ func generateForOneOf(code *jen.File, g *scope.OneOfGroup) {
 				for _, o := range g.Options {
 					code.
 						Line().
-						Id(paramPrefix + o.DiscriminatorFieldName).
+						Id(oneOfCaseFuncName(o)).
 						Func().
 						Params(
 							o.Field.GoType(),
@@ -86,7 +177,7 @@ func generateForOneOf(code *jen.File, g *scope.OneOfGroup) {
 										Id(o.DiscriminatorTypeName),
 								).
 								Return().
-								Id(paramPrefix + o.DiscriminatorFieldName).
+								Id(oneOfCaseFuncName(o)).
 								Call(
 									jen.
 										Id("v").
@@ -100,7 +191,7 @@ func generateForOneOf(code *jen.File, g *scope.OneOfGroup) {
 								jen.Lit(
 									fmt.Sprintf(
 										"%s: x.%s is nil",
-										name,
+										funcName,
 										g.GoFieldName,
 									),
 								),