+++ internal/test/exhaustiveswitch/enum_test.go
@@ -6,7 +6,30 @@ import (
 	. "github.com/dogmatiq/primo/internal/test/exhaustiveswitch"
 )
 
-func TestEnum(t *testing.T) {
+func TestEnumSwitch(t *testing.T) {
+	t.Parallel()
+
+	t.Run(
+		"it calls the function associated with the enum value",
+		func(t *testing.T) {
+			t.Parallel()
+
+			called := false
+
+			Switch_Direction(
+				Direction_LEFT,
+				func(Direction_UNKNOWN_Case) { panic("unexpected UNKNOWN direction") },
+				func(Direction_SINISTER_Case) { called = true },
+				func(Direction_DEXTER_Case) { panic("unexpected LEFT direction") },
+			)
+
+			if !called {
+				t.Fatalf("expected case function to be called")
+			}
+		},
+	)
+}
+func TestEnumMap(t *testing.T) {
 	t.Parallel()
 
 	t.Run(
@@ -15,7 +38,7 @@ func TestEnum(t *testing.T) {
 			t.Parallel()
 
 			want := "<value>"
-			got := Switch_Direction(
+			got := Map_Direction(
 				Direction_LEFT,
 				func(Direction_UNKNOWN_Case) string { panic("unexpected UNKNOWN direction") },
 				func(Direction_SINISTER_Case) string { return "<value>" },